:art: Rename fetchAPI to api.tsx
